+++ lib/rollout/railtie.rb
@@ -1,0 +1,9 @@
+require 'rollout/context_provider'
+
+module Rollout
+  class Railtie < Rails::Railtie
+    initializer 'rollout.controller_extension' do
+      ActionController::Base.send(:include, Rollout::ContextProvider)
+    end
+  end
+end
