+++ lib/rollout/version.rb
@@ -1,3 +1,3 @@
-class Rollout
+module Rollout
   VERSION = "2.0.0c"
 end
